requirements.txt
line-bot-sdk
flask
janome
wordcloud
matplotlib